Update App.js
Exercise with TouchableOpacity
Import TouchableOpacity
Remove Button from view
Add TouchableOpacity
Style it as a button
Add onPressHandler to make it change state from not submitted to submitted and back
Add activeOpacity prop to change it to 0.5. By default it is 0.2. The closer to one, less opaque.
Add Text to button with conditional to show submit if button hasn't been pressed, or show clear when it's been pressed.
Create styling for button.